Refactor/directory (#224)
* Refactoring 예시

* Comment,User axiosRequest 리팩토링

* axiosRequest 마무리

* eslint로 unused-import 삭제 설정
