+++ client/.eslintrc.json
@@ -26,5 +26,6 @@
     "react/prop-types": "off",
     "import/no-named-as-default": 0,
     "import/export": 0
-  }
+  },
+  "plugins": ["unused-imports"]
 }
