+++ client/src/api/Answer/delete.js
@@ -1,0 +1,9 @@
+import axiosInstance from '../../utils/axiosInstance';
+export async function deleteAnswer(answerId) {
+  try {
+    const res = await axiosInstance.delete(`/answers/${answerId}`);
+    alert('삭제되었습니다');
+  } catch (err) {
+    alert(err.response.data.message);
+  }
+}
