+++ client/src/api/Answer/index.js
@@ -1,1 +1,4 @@
 export * from './post';
+export * from './put';
+export * from './delete';
+export * from './votes';
